requirements.txt
torch
transformers
accelerate
qdrant-client
speechrecognition
pyttsx3
openai
ollama
langchain
pyaudio
requests
python-dotenv
gtts
opencv-python
sentence-transformers